Improve todo routes and tests

+++ app/routes.py
@@ -1,4 +1,4 @@
-from flask import render_template, request, redirect, url_for
+from flask import render_template, request, redirect, url_for, abort
 from . import create_app, db
 from .models import Todo
 
@@ -21,7 +21,9 @@ def add_todo():
 
 @app.route('/edit/<int:todo_id>', methods=['POST'])
 def edit_todo(todo_id):
-    todo = Todo.query.get_or_404(todo_id)
+    todo = db.session.get(Todo, todo_id)
+    if not todo:
+        abort(404)
     title = request.form.get('title')
     if title:
         todo.title = title
@@ -30,14 +32,18 @@ def edit_todo(todo_id):
 
 @app.route('/toggle/<int:todo_id>')
 def toggle_todo(todo_id):
-    todo = Todo.query.get_or_404(todo_id)
+    todo = db.session.get(Todo, todo_id)
+    if not todo:
+        abort(404)
     todo.completed = not todo.completed
     db.session.commit()
     return redirect(url_for('index'))
 
 @app.route('/delete/<int:todo_id>')
 def delete_todo(todo_id):
-    todo = Todo.query.get_or_404(todo_id)
+    todo = db.session.get(Todo, todo_id)
+    if not todo:
+        abort(404)
     db.session.delete(todo)
     db.session.commit()
     return redirect(url_for('index'))
